modify ACCUMULATE macro to add by i offset as well
change adam return type to double so it can provide the last seed value in the set of seeds for the previous generated buffer, allowing easy chaining of calls. Only nonce needs to be provided per call.

redo ROUNDS number based on bigger keyspace of 2 ^ 256, and provide more detailed description of what's going on