the heart of the magic happens here. accumulate now accumulates seed content as well as buffer content. here's a brief explanation
there are 18 rounds of the chaotic function applied to the input vector. each round starts with four seeds within (0.0, 0.5). Previous versions of ADAM's RNG process just used the results from the last iteration as the starting seeds for the next, but this is wrong because each round's set of 4 starting seeds shouldn't have any kind of correlation

so now, the starting seeds for all 18 rounds (18 * 4 = 72 seeds in total) are generated in the first step and that array is then passed to apply where it is read at the beginning of each new round. this has really helped the randomness quality