and we're back!
the lack of commits recently is because I have completely switched systems to a M2 Macbook Pro. As such this program needs to be adapted for Apple Silicon, and since there's no scalar routine anymore, that meant learning about ARM NEON and writing adaptations of the old functions that used AVX/AVX-512 to work with ARM chips and consequently Apple Silicon.

So this has been arduous over the past couple weeks, especially in the midst of other life and development stuff, but we are back on the slow track toward regular commits again. Need to do more analysis between the NEON and Intel implementations to make sure they produce exactly the same results and that statistical quality and performance meet all standards. While going through with NEON some enhancements have been made to certain parts of the algorithm which are being tested as well

i'm going to start committing what I can so far though, but expect more complicated stuff later

here you can see the gitignore has been modified to ignore the .DS_STORE MacOS file and to ignore the original cryptography paper itself which I have now placed in the repo for ease of access for myself. This is a paid paper though so I can't distribute it with the repo itself

and the Makefile has been changed to replace \e with \033 for greater portability (\e is not universal) and to add the right SIMD flags for ARM