after taking time to really test and experiment, I'm ready to begin commiting the different changes I've made that are both non-breaking and highly beneficial for statistical quality. The first is reintroducing the RNGSTEP macro logic from ISAAC64, after getting rid of it from before due to not having a use for it really (it felt redundant to have in the diffuse stage) so now I have repurposed it and used it as the method of reseeding the generator! So far the test results prove that this was a good decision, and I will upload more as I compile them
Reseeding is done via passing the RESEED macro

as a result this means the function prototype has changed as well, as the reseeding now happens internally in adam.c and no longer needs to be done by the caller. Notice how seed and nonce are now pointers instead - that's why.