data is initialized in main and modified based on the caller, but most values are default instantiated. but if the user triggers the -b option for example, then reseed is turned on and etc.
durations is currently an array while benchmarking and speeds are being tested. It will revert back to a single double value once I make this thing faster - all the floating point multiplication is killing the performance

a new change which really helped the security of the output sequence was using a 256-bit key now instead of a 64-bit one. Outputing this seed for reuse is currently being figured out (basically how to provide a clean and convenient API for it) so currently -s option doesn't work