add all new necessary imports and update paths for source headers
generate the version string in its own function now instead of just globally in the header

generate the version string for help() the same way

add set_width and set_results functions to handle changing the integer width and number of results returned. You'll see once the new main.c is committed, but basically we don't do any major processing in main anymore. I tidyed up main() and now it just matches options and executes function calls. This is to keep the interface more organized and clean, easier to find stuff. For example, anything related to rng_cli operations and configurations should for the most part be in here (there's a few one liners in main like setting the fmt string, but besides little things like that)