pass rng_data into cli_init to handle full initialization of rng_cli within that method
remove adam -l and the infinite function - this is a painful decision but due to the difficulty in coming up with a cross-terminal solution, the only other option possibly was gtk, but bringing in a GUI dependency to a RNG just felt like overkill/out of place. so perhaps in the future I can make a little widget or something that uses this generator, and people can download it and use it if theyd like. But I believe that ultimately it's not the best thing for this generator. on the bright side, it reduces the code size, binary size, and complexity, and we avoid having to bring in any external dependencies