brand new interface methods because now adam should be able to be compiled as a library that users can incorporate into their applications or tests or whatever. so as of now this is the latest and greatest interface for interacting with ADAM. some major changes include the introduction of double and floating point number generation, which will begin to be introduced over time, and methods to retrieve values of variable width or fill a user provided container of choice. These methods also accept pointers to a double to track how long they take.
The buffer field was removed from the rng_data struct because it is of size 768 per the algorithm. this means that there's no way to enforce the BUF_SIZE limit if the user is given access to the raw pointer, which means they can accidentally go past this limit and just receive random junk whose randomness cant be verified. So now users don't have to worry about index tracking or anything annoying like that