okay so refactor and update the library API stuff into two different types of APIs: a regular one, called the standard API, which is the API as it existed until now, and the new minimal API, which is just one function and gives you a bunch of numbers without anything else.
this is to accomodate different types of uses. there might be someone who is using this who has no need for anything related to doubles, or different widths of integers. the minimal API also seems like it can give way to some FFI action due to how simple and generic it is. In the future, it might be cool to let the user automatically build for rust or go during the build process.

rename adam.c/.h to api.c/.h