adjust the rounds and beta values. BETA will be increased because going from 10E15 to 10E18 allows us to include 3 more digits in the double before casting to u64 and XORing with whatever is in the buffer. It's just used here to pad the u64 with extra bits by adding 3 more digits minimum to the number, so we can safely do this
adjust ROUNDS accordingly. after some more testing it seems like going back to the original algorithm's construction is best. I don't really have the background to make such major changes to the algo if I can't back up their affects on security. On the plus side, the bump in performance hasn't been all that much because I'm not following the algorithm perfectly, but that's a tradeoff I mention in the README.

Also on the bright side, going backto the original algorithm means 24 seeds again instead of 32, so one less chaotic iteration