implement new reseed() function that just houses the latest logic for reseeding adam with new key and nonce. Uses the ISAAC_IND macro to seek to some random byte position in the 3rd map and XOR with the resulting int
The stuff that's commented out was another approach I was experimenting with. Due to the high amount of chaos in the system, in addition to the algorithm being chiefly sensitive to the input vector  (even moreso than the 256-bit seed), I was working with just incrementing the seed's 256-bit value. It didn't seem to get as great of results as this approach, but they both have a negligible effect on overall security - no failures, just some more "unusual" reads with the increment approach. Problem is, the "unusual"'s don't happen consistently or predictably enough for me to know if it's the randomness of the generator or some actual defect. Ironic isn't it? Like suffering from success.