move details about SAC test to test.h header with other tests' info
move ent.h include directly to support.c

give get_seq_properties more responsibility and consolidate examination handling in support.c. cli.c should just be the trigger. so any logic related to the results and testing of these binary sequences should just be kept in support. this allows us to keep cli.c even more tidy and avoid including all of the test.h in its namespace when it doesnt even need to really use it. it needs the testing limits but those have been moved to support.h because they are not used in test.c at all and are used by support.c for validation. so now the examine function is just what it's supposed to be - an interface

also add the floating point permutation printing and calculation logic