create new internal struct for multi-threading called adam_worker. this is so we can properly implement the SAC test, the ENTIRE input must only change by 1 bit, before we were just incrementing the seed but allowing for any nonce which doesn't result in a proper test
so now, we spawn a thread after incrementing the nonce of the main worker and then record the hamming distances after joining the threads

this has required some changes for the function signatures too, namely the ability to work with an arbitrary buffer. This change will probably flood over to the API at some point, but not right now.

accordingly update test.h - remove buffer and chseed properties from rng_test since we use workers now, and update the signature for adam_test to adam_examine which also lets us pass the seed and nonce as this method handles the allocation and cleanup of the workers.

adam_worker is supposed to just be an internal class