this is the latest version of the API consolidated into one format. It uses an opaque struct pointer adam_data to reference a struct which is only defined in api.c, meaning we can ensure that users cannot mess with the state of the RNG in the midst of using it.
Also introduce a couple of new functions that address issues from before like streaming an arbitrary amount of bits to stdout or a file, and the functions for alloc/dealloc of the adam_data struct

Add some more ADAM buffer size units

Finally, don't return the raw pointer to the buffer. Because this can allow users can modify the contents. So we make it const to make sure it is read-only