since we need to allocate memory and deal with files, we add 2 new includes: stdlib and stdio
define the adam_data_s struct which is what adam_data is an alias for. We keep the defined members in this source file so that adam_data can be an opaque pointer struct

finally, change buff_idx to u16 and treat the buffer as bytes not u64, as this allows us to avoid shifting and setting each place to 0 when trying to return numbers by width. This isn't done internally and is bad for the state of the RNG, so using buff_idx as an index to a series of bytes and copying the requested length is much cleaner. this index is reset to 0 upon regeneration