finally implement p-value calculation logic. Seemed impossible for a while and couldn't figure it out, but after some determination I was able to find a good reasonably efficient way to do it given the scope of what I'm trying to do. I pre calculate the gamma values since they take a fixed known parameter (the degrees of freedom / 2) and keep them in test.h, and then in support.c I implemented a modest version of igamc, which calculates the numerator of the chi-square cdf, since gamma is the denominator.
I confirmed using Wolfram Alpha that my igamc implementation returns the right values, so after dividing it by gamma, we just need to subtract from 1 and we get a p-value!

I still report the raw chi-square value, since sometimes the p-values returned are 1 if the results are really not statistically significant at all. By reporting the raw chi-square and conveying the critical value as well, it can let people know whether the test was really passed in case the p-value seems a little wonky, but it shouldn't for the most part as its accurate