make tbt_array just a member of the file rather than the rng_test struct.
implement Walsh-Hadamard randomness test. wh_transform applies the transform to the current sequence, wh_test calculates the sum of all squared test statistics as per the decision options in the paper, and then a p-value is computed accordingly and added to the Fisher's method accumulator. pass rates and distribution of p-values are tracked as well