accuracy issues with larger sequences and calculating p-values for them, so we need to change the statistic size to 128-bits. So we need to do a meta-analysis of meta-analyses, by using Fisher's Method PER 1MB like the Maurer test, in addition to doing it WITHIN each 1MB PER output sequence. So some nested Fisher method action
update header defs for test accordingly

Reason for the change is after double checking the paper, the test statistic calculation can only be approximated by the normal distribution for m >= 7, but in this case it was 5 since our work unit was 32 bits. So to avoid any kind of accuracy or calculation pitfalls, since I can't find an implementation of this test anywhere and am basically a legally blind man with a flashlight in a cave, I want to stick more closely to the conditions laid out in the paper. So now the work unit is the minimum required of 128-bits, providing wins in both accuracy and minimization of p-value deterioration