add 32-bit test - the VNR test. this test also requires z-score table lookup, so a po_zscore method with that logic has been created.
update WHT calculations for accuracy and p-value scaling

now the test suite is finished logically, but needs a bit of a makeover organization wise. but the analytical offerings are finished for now, and we consider not only the original intended range of 4-bit to 64-bit tests, but an 128-bit test as well thanks to me misreading a previous paper. Somehow resulted in a win I guess :P

VNR test also has same p-value scaling at Maurer because it analyzes the same length of sequences