Recently, I went back to the algorithm in the paper. I made some slight deviations from the paper originally to work with the way I structured the SIMD operations, but now I became curious as to how I could optimize further while going back to the original nature of the algorithm, which didn't just XOR the chaotic function results, but used them as indices to permute the existing numbers. The results....
Have been a pretty good speedup, from before where ADAM could produce 1 GB at around 0.45-0.5 seconds, now it successfully can do so at around 0.27-0.37 seconds, although I'm not sure where the tenth of a second discrepancy is coming from. The output also passes the desired cryptographic and statistical tests

This required some changing of the adam_data struct as well, and how api.c uses the primitives. Now accumulate() and diffuse() are used only in the adam_data "setup stage" (aka the adam_setup() method), while the rest of the algorithm is regenerated based on the chaotic function and mixing the output contents. Much more similar to the original algorithm, and I like how we rely a lot more on the actual chaotic function with this approach, leading to better ascertainment of its efficacy as a real cryptographic solution