change chaotic_rsl struct member name
update initialization logic to diffuse the work AND chaotic buffers to increase the variability present in the initial conditions for the RNG. Before only the work buffer (aka output vector) was being mixed with the ISAAC_MIX macro, which can diffuse 512 bits at once, while, while the chaotic maps were all being MEMSET to a random byte.

Testing shows this doesn't seem to be sufficient, so to bolster the scheme we can sacrifice some runtime due to the huge perf gains from 0.9.0 to 0.10.0, and diffuse all of our buffers some more by calling diffuse() on all 3 buffers.

This doesn't really even add up to much of a performance cost anyway, and some suspect seeds that failed some adam -e tests now pass those tests. The rest of the scheme is unchanged from 0.10.0, but this initial complexity produces stronger results